[lld] Introduce result caching
Result caching is enabled using flag `--fcas-cache-results`.

When caching is enabled:
1. The `link` function is called in "dependency scanning" mode, and all
the input file accesses are recorded using a `CachingOnDiskFileSystem`
2. The input files are ingested into a CAS tree along with the arguments
in order to form the caching key.
3. If the result is not cached, the `link` function is called again to
perform its work but with CAS filesystem for the input files that were
collected in the first pass. This ensures that no input file gets used
that was not accounted for as part of the caching key.