Initialize the Swift CompilerInstance with the triple from the load commands.
The serialized triple is the triple of the last binary __swiftast
section that was processed. Instead of relying on the section contents
order, we overwrite the triple in the CompilerInvocation with the
triple recovered from the binary.

(cherry picked from commit ca320d6a636bb81c1173314e857bb86386a6fdcd)