Fix dynamic import of libraries in Xcode playgrounds.
The Swift expression parser uses the fixit mechanism to force
SwiftASTContext to be reinitialized if it contains fatal
errors. Playgrounds disable fixit application and thus defeated the
mechanism. This became a problem after the late dylib notification was
fixed in 1348ae3ce9dc79fd601a7283e6658fb19c27cdb6. This patch detects
the situation where the fixit is identical to the original expression
and makes one retry even if fixits are disabled.

rdar://85431564