WIP: ADT: Add AtomicUniquePointer<T> and LazyAtomicPointer<T>
Add two classes to make it easier to deal with pointers atomically.

`AtomicUniquePointer<T>` adapts `std::atomic<T*>` to add ownership
semantics like `std::unique_ptr<T>`. Most APIs have obvious
translations, but it wasn't obvious how to implement
`compare_exchange_{weak,strong}`. In this case, I made the "existing"
value `T*&` and the "new" value `std::unique<T>&`, the latter of which
takes ownership for the old/existing value on success.

`LazyAtomicPointer<T>` has the semantics of `std::atomic<T*>` but uses
`std::atomic<uintptr_t>` as the storage. The major change is that it
adds `T &loadOrGenerate(function_ref<T *()>)`, is acts like the pattern
LLVM usually calls `getOrCreate()`: if there's already a value, it
returns it; else, it takes exclusive access (using `-1` as a sentinel),
generates a value and stores it. Other APIs need minor adjustments to
ensure consistent behaviour if `-1` is seen. If possible, they return
as if `nullptr`; otherwise, they spin, waiting for the generating thread
to store the value.