Prepare TypeRefTypesystem to read static size info from Debug Info
This patch refactors GetBitSize and GetTypeBitAlign to make it easy
let it read size info for static types from debug info. This would be
faster than reading it out of reflection metadata, but the Swift
compiler currently wrongly emit static size info for types that
shouldn't have any, including resilient types. We either need to stop
the compiler from emitting it or we need to mark resilient types to
tell LLDB that the size info cannot be trusted.