[debug-info] If one sees a spill with a dbg.addr use, salvageDebugInfo upon it and don't hoist it.
This ensures that if we have a dbg.addr in a coroutine funclet that is on one of
our function arguments, that the dbg.addr is not mapped to undef and also that
later it isn't hoisted to the front of the basic block. Instead it remains at
its original cloned location.

rdar://83957028

Differential Revision: https://reviews.llvm.org/D119576

(cherry picked from commit 19279ffc77b8d224c447d4eb0ee0c727ab64babf)

Conflicts:
	llvm/lib/Transforms/Coroutines/CoroFrame.cpp

The conflict was that the upstream code here seems to handle more cases. I just
did the same transform I did in OSS here (which is not hoisting llvm.dbg.addr).
The rest of the code is unchanged.