CAS: Update CASID from ArrayRef<uint8_t> => internal CAS ID
Previously, CASID wrapped an `ArrayRef<uint8_t>`. This was a bit slow,
since the CAS needed a fresh lookup on every access. It also always felt
dangerous... what's the lifetime of the memory? who owns it? and when is
it safe to compare two IDs?

Now, the CASID wraps an ID that's internal to a CAS instance. This is
should be possible for any CAS implementation to support. Even if there
is no natural ID to return, the hashes can be stored in an in-memory map
for the instance to generate such IDs, but it's unlikely this will
require extra storage. If the CAS has in-memory data structures, a
pointer to the in-memory handle should work. If the CAS is on-disk, a
file offset should work.

This internal ID is `uint64_t` (not `uintptr_t`) to allow a CAS
implemention running on a system with 32-bit pointers to use IDs big
enough for large data sets.

We still want to compare CASIDs between different CAS instances; at
least, this is useful for testing; likely it'd be important for some use
cases in practice too. To facilitate that, there's a new class
CASIDContext (a new base class of CASDB) with two virtual methods:

- `StringRef getHashSchemaIdentifier()`, which returns a unique
  identifier for the hashing schema, which should be the same across all
  CAS instances whose CASIDs are compatible (i.e., that have the same
  schema for how to hash objects and use the same hash function).
- `ArrayRef<uint8_t> getHashImpl(const CASID &)`, an implementation
  method for `CASID::getHash()`. That API is still important for things
  like `DenseMapInfo`, which would want CASIDs from different
  CASIDContext instances to hash the same way if they have the same
  hash.

If two CASIDs have the same CASIDContext, then the internal IDs can be
compared directly; otherwise, if they have the same
getHashSchemaIdentifier(), the hashes are compared; otherwise, they're
considered unequal.

Two follow-ups to do later:

- Code that is collecting/mapping/etc. lots of CASIDs that are
  guaranteed to be from the same CASIDContext can extract the internal
  ID (using CASID::getInternalID()) to optimize storage. I'm thinking of
  the CAS-based filesystems and some builder data structures in
  CASObjectFormats.
- BlobRef/TreeRef/NodeRef should probably get similar treatment and be
  stripped down to a simple object handle. I think we should do a
  wrapper around a `uint64_t` value; it'd be up to the CAS
  implementation whether this was the same value as the internal ID in
  CASID or something "closer" to the data.