[lldb] Determine the main binary in JSON crashlogs
The symbolicator assumes that the first image in the image list is the
main image. That isn't always the case. For JSON crashlogs we can use
the procName to move the main image to the front of the list.

rdar://83907760
(cherry picked from commit 0f29319e56459eb26a3bf7fb86e45e63f43a168e)