Avoid initializing SwiftASTContext when looking up Clang names. (NFC)
With this change we can set a breakpoint on
SwiftASTContext::SwiftASTContext() in a non-asserts LLDB and run
TestSwiftFoundationTypeURL and the breakpoint will not trigger.  We do
this by initializing an empty swift::ASTContext, whose sole purpose is
to host a a ClangImporter's imported DeclNames. Initializing it is
safe & fast, since no modules are being imported.