[lldb] Always move the cursor back after printing progress
This got lost while iterating on the patch. We need to always move the
cursor to the front of the line so that if something else
(asynchronously) prints to the debugger's output it overwrites the
progress message.

(cherry picked from commit 116715270d07d18d101ff8d1b4a42cfd76c5cc55)