[debug-info] Debug salvage llvm.dbg.addr in original function that point into the coroutine frame when splitting coros.
We are already doing this in the split functions while we clone. This just
handles the original function.

I also updated the coroutine split test to validate that we are always referring
to the msg in the context object instead of in a shadow copy.

rdar://83957028

Reviewed By: aprantl

Differential Revision: https://reviews.llvm.org/D121324

(cherry picked from commit 0b647fc5299156bf83c46aa539d6c9c39647bb36)

Conflicts:
	llvm/lib/Transforms/Coroutines/CoroSplit.cpp
	llvm/test/Transforms/Coroutines/coro-debug-frame-variable.ll