CAS.o: Add -object-stats-format=<csv, pretty> and an argument to -object-stats
Add `-object-stats-format=csv` to llvm-cas-object-format. This changes the
output format of `-object-stats` to comma-separated.

`-object-stats-format=pretty` will use the standard output that `-object-stats` used to use

As a drive-by, change `-object-stats` to require an argument for where to
send the stats. This allows the stats to be saved separately from the other
output.