CAS: Fix a too-narrow assertion in the OnDiskCAS
Fix an assertion about null-termination in the OnDiskCAS that was too
narrow.

The assertion was checking that `MemoryBufferRef` was null-terminated.
This was the right check for the `TrieRecordData::Blob*` on-disk
formats, but not for `TrieRecordData::Standalone`. Non-blobs do not care
if the `mmap` itself is null-terminated, but do care if the "data"
section is null-terminated.

This splits the assertion in two.

The unit tests add tests for blobs (covering Blob and Blob0 formats) and
nodes (covering Standalone). Only the test for nodes was failing before
applying the fix.

This patch is instead of: https://github.com/apple/llvm-project/pull/4059