[lldb/Process] Silence toolchain mismatch warnings for Scripted Processes (NFC)
When using Scripted Processes, it's very likely that lldb's version
might not align with the swift compiler that built the target.

Thus, this will cause the toolchain mismatch warnings to be shown to the
user every time they "stop" in a Swift frame.

Since we shouldn't need Swift's AST to symbolicate the stopped stackframe,
we can safely silence these warnings for Scripted Processes.

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>