[c++-interop] Test that APINotes handles ObjCMethods while in C++-Interop
In a previous commit I taught APINotes to treat extern-c context as a
toplevel file context, but this only makes sense when we are handling
the decls handled in the block like VarDecls, FunctionDecls,
ObjCInterfaceDecls, etc.

But when the decl being handled for APINote is an ObjCMethod inside of a
ObjCContainer then it is important to allow the subsequent code in the
Sema::ProcessAPINotes function to handle things.

This can be excercised by the following APINote:

```
---
Name: SomeModule
Classes:
- Name: NSSomeClass
  SwiftName: SomeClass
  Methods:
  - Selector: 'didMoveToParentViewController:'
    SwiftName: didMove(toParent:)
    MethodKind: Instance
```

This commit is now just a test case due to another commit that has
already landed.