fdepscan: Avoid (ignored) -triple -cc1 in -fdepscan=daemon response file, mostly NFC
The old logic for parsing args sent back from the daemon, duplicated in
the last refactoring commit, started the response file with `"-triple"
"-cc1"`. This had no effect since there was always a `"-triple"` later
on to override it, but it's a bit silly.

Now the -fdepscan=daemon response file starts with `"-cc1"`, just like
-fdepscan=inline. This is itself redundant since the command that uses
it calls `-cc1 @rsp`, but it works and (unlike before) it doesn't *look*
broken.